feat(ga): remove reduntant file
